crud de todos os objetos atualizada

+++ src/main/java/com/example/demo/repositories/PlayerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayerRepository extends JpaRepository<Player, Long> {
-    // O Spring cria o CRUD básico automaticamente aqui dentro
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
+    // CRUD básico estruturado para os jogadores (usando o ID Integer do seu script)
 }
